[API] Refactor by adding services for controllers
