+++ src/main/java/com/dezso/varga/backgammon/authentication/AuthenticationController.java
@@ -1,12 +1,7 @@
 package com.dezso.varga.backgammon.authentication;
 
-import com.dezso.varga.backgammon.authentication.domain.Account;
-import com.dezso.varga.backgammon.authentication.repository.AccountRepository;
-import com.dezso.varga.backgammon.exeptions.AuthExeption;
-import com.dezso.varga.backgammon.exeptions.BgException;
-import com.dezso.varga.backgammon.exeptions.UserAlreadyExistsException;
+import com.dezso.varga.backgammon.authentication.services.AuthenticationService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 import com.dezso.varga.backgammon.authentication.domain.RegisterRequest;
 
@@ -16,56 +11,22 @@ public class AuthenticationController {
 	
 	@Autowired
 	private AuthenticationService authenticationService;
-	@Autowired
-	private AccountRepository accountRepository;
 
 	@RequestMapping(method=RequestMethod.POST,value="/register")
 	public String signup(@RequestBody RegisterRequest registerRequest) throws Exception{
-		if (registerRequest == null
-				|| registerRequest.getAccount().getFirstName() == null
-				|| registerRequest.getAccount().getFirstName().trim().isEmpty()
-				|| registerRequest.getAccount().getLastName() == null
-				|| registerRequest.getAccount().getLastName().trim().isEmpty()
-				|| registerRequest.getAccount().getEmail() == null
-				|| registerRequest.getAccount().getEmail().trim().isEmpty()
-				|| registerRequest.getAccount().getPassword() == null
-				|| registerRequest.getAccount().getPassword().trim().isEmpty()) {
-			throw new BgException("Missing or invalid mandatory fields at registration",
-					HttpStatus.PRECONDITION_FAILED.value());
-		}
-		Account existingAccount = accountRepository.findByEmail(registerRequest.getAccount().getEmail().trim());
-		if (existingAccount != null) {
-			throw new UserAlreadyExistsException("Account already exists", HttpStatus.CONFLICT.value());
-		}
-		String confirmationToken = AuthUtils.generateRegisterConfirmationToken(registerRequest);
 		authenticationService.sendConfirmationMail();
-		return confirmationToken;
+		return authenticationService.getConfirmationToken(registerRequest);
 	}
 
 	@RequestMapping(method=RequestMethod.GET, value="register/confirm")
 	public String confirm(@RequestHeader (value="Authorization") String confirmToken) throws Exception {
-		Account account = AuthUtils.validateConfirmToken(confirmToken);
-		Account existingAccount = accountRepository.findByEmail(account.getEmail());
-		if (existingAccount == null) {
-			accountRepository.save(account);
-		} else {
-			throw new UserAlreadyExistsException("User already verified", HttpStatus.CONFLICT.value());
-		}
 
-		return AuthUtils.generateBearerToken(account);
+		return AuthUtils.generateBearerToken(authenticationService.saveAccount(confirmToken));
 	}
 
 	@RequestMapping(method=RequestMethod.POST, value="/login")
 	public String login(@RequestHeader (value="Authorization", required=false) String authHeader) throws Exception {
-		Account credentials = AuthUtils.extractAccountFromBasicToken(authHeader);
-		Account account = accountRepository.findByEmail(credentials.getEmail());
-		if (account == null) {
-			throw new AuthExeption("User email not found.", HttpStatus.UNAUTHORIZED.value());
-		}
-		if (!credentials.getPassword().equals(account.getPassword())) {
-			throw new AuthExeption("Invalid credentials. Please check your email and password.",
-					HttpStatus.UNAUTHORIZED.value());
-		}
-		return AuthUtils.generateBearerToken(account);
+
+		return AuthUtils.generateBearerToken(authenticationService.login(authHeader));
 	}
 }
