+++ src/main/java/com/dezso/varga/backgammon/home/HomeController.java
@@ -1,13 +1,11 @@
 package com.dezso.varga.backgammon.home;
 
 import com.dezso.varga.backgammon.authentication.domain.Account;
-import com.dezso.varga.backgammon.authentication.repository.AccountRepository;
+import com.dezso.varga.backgammon.authentication.services.AccountService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -18,13 +16,10 @@ import java.util.List;
 public class HomeController {
 
     @Autowired
-    private AccountRepository accountRepository;
+    private AccountService accountService;
 
     @RequestMapping(method= RequestMethod.GET, value="/users")
     public List<Account> getAccounts() throws Exception {
-        List<Account> users = new ArrayList<>();
-        accountRepository.findAll().forEach(users::add);
-
-        return users;
+        return accountService.findAllAccounts();
     }
 }
